Rename all file names to lowercase_with_underscore
