+++ lib/PDFConversion.dart
@@ -12,8 +12,8 @@ import 'package:pdf/widgets.dart' as pw;
 import 'package:open_file/open_file.dart';
 // import 'package:shared_preferences/shared_preferences.dart';
 
-import 'PDFPreviewScreen.dart';
-import 'Providers/ImageList.dart';
+import 'pdf_preview_screen.dart';
+import 'Providers/image_list.dart';
 
 class PDFConversion extends StatefulWidget {
   final ImageList list;
